requirements.txt
# MCP Framework
mcp>=1.0.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0

# Async Support
asyncio-throttle>=1.0.2

# Logging
structlog>=23.2.0

# Configuration
python-dotenv>=1.0.0

# Date/Time Handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
black>=23.9.0
isort>=5.12.0
mypy>=1.6.0

# Monitoring
psutil>=5.9.0

# Optional: Redis for caching (uncomment if needed)
# redis>=5.0.0
# aioredis>=2.0.0